feature(nemesis.py): added into _add_and_init_new_cluster_node
function a random call to new Scylla feature
(disable/enable autocompaction), so during the bootstrap
phase, if calling the command, the autocompaction
will be disabled. Per dev explanation, it is a feature
to be used during DR, and we want just to be sure this
feature isn't broken at any point.
At the end of bootstrap, we enable autocompaction to
have the cluster running regularly.